+++ frontend/src/components/utils/buttons/circle-button.tsx
@@ -4,6 +4,7 @@ import { HTMLAttributes } from 'react';
 const CircleButton: React.FunctionComponent<HTMLAttributes<HTMLButtonElement>> = (props) => {
   return (
     <button
+      type="button"
       {...props}
       className={classNames(
         `rounded-full border-2 border-gray-400 secondary px-4 py-1`,